refactor: Remove global context handling from api code
