+++ .env.template
@@ -74,4 +74,6 @@ LITELLM_LOG="ERROR"
 #       Relational: SQLite, Postgres
 #       Vector: LanceDB
 #       Graph: KuzuDB
+#
+# It enforces LanceDB and KuzuDB use and uses them to create databases per Cognee user + dataset
 ENABLE_BACKEND_ACCESS_CONTROL=False
